requirements.txt
flask
requests
azure-ai-textanalytics
